cleanup viz structure
