+++ tidy3d/components/viz/__init__.py
@@ -11,7 +11,6 @@ import pydantic.v1 as pd
 try:
     import matplotlib.pyplot as plt
     import matplotlib.ticker as ticker
-    from matplotlib import cycler
     from matplotlib.colors import is_color_like
     from matplotlib.patches import ArrowStyle, PathPatch
     from matplotlib.path import Path
@@ -25,7 +24,6 @@ from numpy import array, concatenate, inf, ones
 
 from ...constants import UnitScaling
 from ...exceptions import SetupError, Tidy3dKeyError
-from ...log import log
 from ..base import Tidy3dBaseModel
 from ..types import Ax, Axis, LengthUnit
 
@@ -37,17 +35,6 @@ PLOT_BUFFER = 0.3
 ARROW_ALPHA = 0.8
 ARROW_LENGTH = 0.3
 
-FLEXCOMPUTE_COLORS = {
-    "brand_green": "#00643C",
-    "brand_tan": "#B8A18B",
-    "brand_blue": "#6DB5DD",
-    "brand_purple": "#8851AD",
-    "brand_black": "#000000",
-    "brand_orange": "#FC7A4C",
-}
-ARROW_COLOR_SOURCE = FLEXCOMPUTE_COLORS["brand_green"]
-ARROW_COLOR_POLARIZATION = FLEXCOMPUTE_COLORS["brand_tan"]
-ARROW_COLOR_MONITOR = FLEXCOMPUTE_COLORS["brand_orange"]
 
 """ Decorators """
 
@@ -167,23 +154,6 @@ plot_params_lumped_element = PlotParams(
     alpha=0.4, facecolor="mediumblue", edgecolor="mediumblue", lw=3
 )
 
-# stores color of simulation.structures for given index in simulation.medium_map
-MEDIUM_CMAP = [
-    "#689DBC",
-    "#D0698E",
-    "#5E6EAD",
-    "#C6224E",
-    "#BDB3E2",
-    "#9EC3E0",
-    "#616161",
-    "#877EBC",
-]
-
-# colormap for structure's permittivity in plot_eps
-STRUCTURE_EPS_CMAP = "gist_yarg"
-STRUCTURE_EPS_CMAP_R = "gist_yarg_r"
-STRUCTURE_HEAT_COND_CMAP = "gist_yarg"
-
 
 def is_valid_color(value: str) -> str:
     if not is_color_like(value):
@@ -223,52 +193,6 @@ class VisualizationSpec(Tidy3dBaseModel):
             return is_valid_color(values["facecolor"])
 
         return is_valid_color(value)
-
-
-# Tidy3D default plotting style parameters
-_TIDY3D_STYLE_PARAMS = {
-    "axes.prop_cycle": cycler(
-        color=[
-            FLEXCOMPUTE_COLORS["brand_green"],
-            FLEXCOMPUTE_COLORS["brand_tan"],
-            FLEXCOMPUTE_COLORS["brand_blue"],
-            FLEXCOMPUTE_COLORS["brand_purple"],
-        ]
-    ),
-    "axes.grid": True,
-    "grid.color": "#CDCDCD",
-    "grid.linestyle": ":",
-    "axes.edgecolor": "#ECEBEA",
-}
-
-# Store original parameters before applying Tidy3D style
-_ORIGINAL_PARAMS = {}
-
-try:
-    # Store only the parameters that Tidy3D style will modify
-    for key in _TIDY3D_STYLE_PARAMS:
-        if key in plt.rcParams:
-            _ORIGINAL_PARAMS[key] = plt.rcParams[key]
-    # Apply the Tidy3D style automatically on import
-    plt.rcParams.update(_TIDY3D_STYLE_PARAMS)
-except Exception as e:
-    log.error(f"Failed to apply Tidy3D plotting style on import. Error: {e}")
-    _ORIGINAL_PARAMS = {}  # Clear original params if application failed
-
-
-def reset_previous_style():
-    """
-    Resets matplotlib rcParams to the values they had before the Tidy3D
-    style was automatically applied on import.
-    """
-    if not _ORIGINAL_PARAMS:
-        log.warning("No previous Matplotlib style state found to reset to.")
-        return
-
-    try:
-        plt.rcParams.update(_ORIGINAL_PARAMS)
-    except Exception as e:
-        log.error(f"Failed to reset previous Matplotlib style. Error: {e}")
 
 
 """=================================================================================================
