+++ tidy3d/components/viz/color_paletes.py
@@ -1,0 +1,3304 @@
+SEQUENTIAL_PALETTES_HEX = {
+    "flex_turquoise_seq": [
+        "#ffffff",
+        "#fefefe",
+        "#fdfdfd",
+        "#fcfcfc",
+        "#fbfbfb",
+        "#fafbfa",
+        "#f9fafa",
+        "#f8f9f9",
+        "#f7f8f8",
+        "#f6f7f7",
+        "#f5f6f6",
+        "#f3f5f5",
+        "#f2f4f4",
+        "#f1f3f3",
+        "#f0f3f2",
+        "#eff2f1",
+        "#eef1f1",
+        "#edf0f0",
+        "#ecefef",
+        "#ebeeee",
+        "#eaeded",
+        "#e9edec",
+        "#e8eceb",
+        "#e7ebeb",
+        "#e6eaea",
+        "#e5e9e9",
+        "#e4e8e8",
+        "#e3e7e7",
+        "#e2e7e6",
+        "#e1e6e5",
+        "#e0e5e5",
+        "#dfe4e4",
+        "#dee3e3",
+        "#dde2e2",
+        "#dce2e1",
+        "#dbe1e0",
+        "#dae0df",
+        "#d9dfdf",
+        "#d8dede",
+        "#d7dedd",
+        "#d6dddc",
+        "#d5dcdb",
+        "#d4dbdb",
+        "#d3dada",
+        "#d2dad9",
+        "#d1d9d8",
+        "#d1d8d7",
+        "#d0d7d6",
+        "#cfd6d6",
+        "#ced6d5",
+        "#cdd5d4",
+        "#ccd4d3",
+        "#cbd3d2",
+        "#cad2d2",
+        "#c9d2d1",
+        "#c8d1d0",
+        "#c7d0cf",
+        "#c6cfce",
+        "#c5cece",
+        "#c4cecd",
+        "#c3cdcc",
+        "#c2cccb",
+        "#c1cbca",
+        "#c0cbca",
+        "#bfcac9",
+        "#bec9c8",
+        "#bec8c7",
+        "#bdc8c7",
+        "#bcc7c6",
+        "#bbc6c5",
+        "#bac5c4",
+        "#b9c5c3",
+        "#b8c4c3",
+        "#b7c3c2",
+        "#b6c2c1",
+        "#b5c2c0",
+        "#b4c1c0",
+        "#b3c0bf",
+        "#b2bfbe",
+        "#b2bfbd",
+        "#b1bebd",
+        "#b0bdbc",
+        "#afbcbb",
+        "#aebcba",
+        "#adbbba",
+        "#acbab9",
+        "#abbab8",
+        "#aab9b7",
+        "#a9b8b7",
+        "#a9b7b6",
+        "#a8b7b5",
+        "#a7b6b4",
+        "#a6b5b4",
+        "#a5b4b3",
+        "#a4b4b2",
+        "#a3b3b2",
+        "#a2b2b1",
+        "#a1b2b0",
+        "#a1b1af",
+        "#a0b0af",
+        "#9fb0ae",
+        "#9eafad",
+        "#9daeac",
+        "#9cadac",
+        "#9badab",
+        "#9aacaa",
+        "#99abaa",
+        "#99aba9",
+        "#98aaa8",
+        "#97a9a7",
+        "#96a9a7",
+        "#95a8a6",
+        "#94a7a5",
+        "#93a6a5",
+        "#92a6a4",
+        "#92a5a3",
+        "#91a4a2",
+        "#90a4a2",
+        "#8fa3a1",
+        "#8ea2a0",
+        "#8da2a0",
+        "#8ca19f",
+        "#8ca09e",
+        "#8ba09e",
+        "#8a9f9d",
+        "#899e9c",
+        "#889e9c",
+        "#879d9b",
+        "#869c9a",
+        "#869c9a",
+        "#859b99",
+        "#849a98",
+        "#839a97",
+        "#829997",
+        "#819896",
+        "#809895",
+        "#809795",
+        "#7f9694",
+        "#7e9693",
+        "#7d9593",
+        "#7c9492",
+        "#7b9491",
+        "#7a9391",
+        "#7a9290",
+        "#79928f",
+        "#78918f",
+        "#77908e",
+        "#76908d",
+        "#758f8d",
+        "#758f8c",
+        "#748e8b",
+        "#738d8b",
+        "#728d8a",
+        "#718c89",
+        "#708b89",
+        "#708b88",
+        "#6f8a87",
+        "#6e8987",
+        "#6d8986",
+        "#6c8885",
+        "#6b8885",
+        "#6a8784",
+        "#6a8684",
+        "#698683",
+        "#688582",
+        "#678482",
+        "#668481",
+        "#658380",
+        "#658280",
+        "#64827f",
+        "#63817e",
+        "#62817e",
+        "#61807d",
+        "#607f7c",
+        "#607f7c",
+        "#5f7e7b",
+        "#5e7d7b",
+        "#5d7d7a",
+        "#5c7c79",
+        "#5b7c79",
+        "#5b7b78",
+        "#5a7a77",
+        "#597a77",
+        "#587976",
+        "#577975",
+        "#567875",
+        "#567774",
+        "#557774",
+        "#547673",
+        "#537572",
+        "#527572",
+        "#517471",
+        "#507470",
+        "#507370",
+        "#4f726f",
+        "#4e726f",
+        "#4d716e",
+        "#4c716d",
+        "#4b706d",
+        "#4b6f6c",
+        "#4a6f6b",
+        "#496e6b",
+        "#486e6a",
+        "#476d6a",
+        "#466c69",
+        "#456c68",
+        "#446b68",
+        "#446b67",
+        "#436a67",
+        "#426966",
+        "#416965",
+        "#406865",
+        "#3f6864",
+        "#3e6763",
+        "#3e6663",
+        "#3d6662",
+        "#3c6562",
+        "#3b6561",
+        "#3a6460",
+        "#396360",
+        "#38635f",
+        "#37625f",
+        "#36625e",
+        "#35615d",
+        "#35605d",
+        "#34605c",
+        "#335f5c",
+        "#325f5b",
+        "#315e5a",
+        "#305d5a",
+        "#2f5d59",
+        "#2e5c58",
+        "#2d5c58",
+        "#2c5b57",
+        "#2b5a57",
+        "#2a5a56",
+        "#295955",
+        "#285955",
+        "#275854",
+        "#265754",
+        "#255753",
+        "#245652",
+        "#235652",
+        "#225551",
+        "#215551",
+        "#205450",
+        "#1e534f",
+        "#1d534f",
+        "#1c524e",
+        "#1b524e",
+        "#1a514d",
+        "#18504c",
+        "#17504c",
+        "#164f4b",
+        "#144f4b",
+        "#134e4a",
+    ],
+    "flex_green_seq": [
+        "#ffffff",
+        "#fefefe",
+        "#fdfdfd",
+        "#fcfcfc",
+        "#fbfbfb",
+        "#f9fafa",
+        "#f8f9f9",
+        "#f7f8f8",
+        "#f6f7f7",
+        "#f5f6f6",
+        "#f4f5f5",
+        "#f3f5f3",
+        "#f2f4f2",
+        "#f1f3f1",
+        "#f0f2f0",
+        "#eff1ef",
+        "#eef0ee",
+        "#ecefed",
+        "#ebeeec",
+        "#eaedeb",
+        "#e9ecea",
+        "#e8ebe9",
+        "#e7eae8",
+        "#e6eae7",
+        "#e5e9e6",
+        "#e4e8e5",
+        "#e3e7e4",
+        "#e2e6e3",
+        "#e1e5e2",
+        "#e0e4e1",
+        "#dfe3e0",
+        "#dee3df",
+        "#dde2de",
+        "#dce1dd",
+        "#dbe0dc",
+        "#dadfdb",
+        "#d8deda",
+        "#d7ddd9",
+        "#d6dcd8",
+        "#d5dcd7",
+        "#d4dbd6",
+        "#d3dad5",
+        "#d2d9d4",
+        "#d1d8d3",
+        "#d0d7d2",
+        "#cfd6d1",
+        "#ced6d0",
+        "#cdd5cf",
+        "#ccd4ce",
+        "#cbd3ce",
+        "#cad2cd",
+        "#c9d1cc",
+        "#c8d1cb",
+        "#c7d0ca",
+        "#c6cfc9",
+        "#c5cec8",
+        "#c4cdc7",
+        "#c3cdc6",
+        "#c2ccc5",
+        "#c1cbc4",
+        "#c0cac3",
+        "#bfc9c2",
+        "#bec9c1",
+        "#bdc8c0",
+        "#bcc7bf",
+        "#bbc6be",
+        "#bac5bd",
+        "#b9c5bd",
+        "#b9c4bc",
+        "#b8c3bb",
+        "#b7c2ba",
+        "#b6c1b9",
+        "#b5c1b8",
+        "#b4c0b7",
+        "#b3bfb6",
+        "#b2beb5",
+        "#b1bdb4",
+        "#b0bdb3",
+        "#afbcb2",
+        "#aebbb1",
+        "#adbab1",
+        "#acbab0",
+        "#abb9af",
+        "#aab8ae",
+        "#a9b7ad",
+        "#a8b7ac",
+        "#a7b6ab",
+        "#a6b5aa",
+        "#a5b4a9",
+        "#a5b4a8",
+        "#a4b3a8",
+        "#a3b2a7",
+        "#a2b1a6",
+        "#a1b1a5",
+        "#a0b0a4",
+        "#9fafa3",
+        "#9eaea2",
+        "#9daea1",
+        "#9cada0",
+        "#9baca0",
+        "#9aab9f",
+        "#99ab9e",
+        "#99aa9d",
+        "#98a99c",
+        "#97a89b",
+        "#96a89a",
+        "#95a799",
+        "#94a699",
+        "#93a598",
+        "#92a597",
+        "#91a496",
+        "#90a395",
+        "#90a394",
+        "#8fa293",
+        "#8ea193",
+        "#8da092",
+        "#8ca091",
+        "#8b9f90",
+        "#8a9e8f",
+        "#899e8e",
+        "#889d8d",
+        "#879c8d",
+        "#879b8c",
+        "#869b8b",
+        "#859a8a",
+        "#849989",
+        "#839988",
+        "#829888",
+        "#819787",
+        "#809786",
+        "#809685",
+        "#7f9584",
+        "#7e9483",
+        "#7d9483",
+        "#7c9382",
+        "#7b9281",
+        "#7a9280",
+        "#79917f",
+        "#79907e",
+        "#78907e",
+        "#778f7d",
+        "#768e7c",
+        "#758e7b",
+        "#748d7a",
+        "#738c79",
+        "#728c79",
+        "#728b78",
+        "#718a77",
+        "#708a76",
+        "#6f8975",
+        "#6e8875",
+        "#6d8774",
+        "#6c8773",
+        "#6c8672",
+        "#6b8571",
+        "#6a8571",
+        "#698470",
+        "#68836f",
+        "#67836e",
+        "#66826d",
+        "#66816d",
+        "#65816c",
+        "#64806b",
+        "#637f6a",
+        "#627f69",
+        "#617e69",
+        "#607d68",
+        "#607d67",
+        "#5f7c66",
+        "#5e7c65",
+        "#5d7b65",
+        "#5c7a64",
+        "#5b7a63",
+        "#5a7962",
+        "#5a7861",
+        "#597861",
+        "#587760",
+        "#57765f",
+        "#56765e",
+        "#55755d",
+        "#55745d",
+        "#54745c",
+        "#53735b",
+        "#52725a",
+        "#51725a",
+        "#507159",
+        "#4f7058",
+        "#4f7057",
+        "#4e6f56",
+        "#4d6e56",
+        "#4c6e55",
+        "#4b6d54",
+        "#4a6d53",
+        "#4a6c53",
+        "#496b52",
+        "#486b51",
+        "#476a50",
+        "#466950",
+        "#45694f",
+        "#44684e",
+        "#44674d",
+        "#43674c",
+        "#42664c",
+        "#41654b",
+        "#40654a",
+        "#3f6449",
+        "#3e6449",
+        "#3e6348",
+        "#3d6247",
+        "#3c6246",
+        "#3b6146",
+        "#3a6045",
+        "#396044",
+        "#385f43",
+        "#385e43",
+        "#375e42",
+        "#365d41",
+        "#355c40",
+        "#345c40",
+        "#335b3f",
+        "#325b3e",
+        "#315a3d",
+        "#30593d",
+        "#30593c",
+        "#2f583b",
+        "#2e573a",
+        "#2d573a",
+        "#2c5639",
+        "#2b5538",
+        "#2a5537",
+        "#295437",
+        "#285436",
+        "#275335",
+        "#265234",
+        "#265234",
+        "#255133",
+        "#245032",
+        "#235031",
+        "#224f31",
+        "#214e30",
+        "#204e2f",
+        "#1f4d2e",
+        "#1e4c2e",
+        "#1d4c2d",
+        "#1c4b2c",
+        "#1b4b2b",
+        "#1a4a2b",
+        "#18492a",
+        "#174929",
+        "#164828",
+        "#154728",
+        "#144727",
+        "#134626",
+        "#124525",
+        "#104525",
+        "#0f4424",
+    ],
+    "flex_blue_seq": [
+        "#ffffff",
+        "#fefefe",
+        "#fdfdfd",
+        "#fbfcfc",
+        "#fafbfb",
+        "#f9fafa",
+        "#f8f9f9",
+        "#f7f7f8",
+        "#f6f6f8",
+        "#f4f5f7",
+        "#f3f4f6",
+        "#f2f3f5",
+        "#f1f2f4",
+        "#f0f1f3",
+        "#eff0f2",
+        "#eeeff1",
+        "#eceef0",
+        "#ebedf0",
+        "#eaecef",
+        "#e9ebee",
+        "#e8eaed",
+        "#e7e9ec",
+        "#e6e8eb",
+        "#e4e7ea",
+        "#e3e6ea",
+        "#e2e5e9",
+        "#e1e4e8",
+        "#e0e3e7",
+        "#dfe2e6",
+        "#dee1e5",
+        "#dde0e5",
+        "#dcdfe4",
+        "#dadee3",
+        "#d9dde2",
+        "#d8dce1",
+        "#d7dbe0",
+        "#d6dae0",
+        "#d5d9df",
+        "#d4d8de",
+        "#d3d7dd",
+        "#d2d6dc",
+        "#d1d5dc",
+        "#d0d4db",
+        "#cfd3da",
+        "#ced2d9",
+        "#ccd1d9",
+        "#cbd0d8",
+        "#cacfd7",
+        "#c9ced6",
+        "#c8cdd5",
+        "#c7ccd5",
+        "#c6ccd4",
+        "#c5cbd3",
+        "#c4cad2",
+        "#c3c9d2",
+        "#c2c8d1",
+        "#c1c7d0",
+        "#c0c6cf",
+        "#bfc5cf",
+        "#bec4ce",
+        "#bdc3cd",
+        "#bcc2cd",
+        "#bbc1cc",
+        "#bac0cb",
+        "#b9c0ca",
+        "#b8bfca",
+        "#b7bec9",
+        "#b6bdc8",
+        "#b5bcc7",
+        "#b4bbc7",
+        "#b3bac6",
+        "#b2b9c5",
+        "#b1b8c5",
+        "#b0b7c4",
+        "#afb7c3",
+        "#aeb6c3",
+        "#adb5c2",
+        "#acb4c1",
+        "#abb3c1",
+        "#aab2c0",
+        "#a9b1bf",
+        "#a8b0be",
+        "#a7b0be",
+        "#a6afbd",
+        "#a5aebc",
+        "#a4adbc",
+        "#a3acbb",
+        "#a2abba",
+        "#a1aaba",
+        "#a0aab9",
+        "#9fa9b8",
+        "#9ea8b8",
+        "#9da7b7",
+        "#9ca6b7",
+        "#9ba5b6",
+        "#9aa4b5",
+        "#99a4b5",
+        "#98a3b4",
+        "#97a2b3",
+        "#96a1b3",
+        "#95a0b2",
+        "#949fb1",
+        "#939fb1",
+        "#929eb0",
+        "#919db0",
+        "#909caf",
+        "#8f9bae",
+        "#8f9aae",
+        "#8e9aad",
+        "#8d99ac",
+        "#8c98ac",
+        "#8b97ab",
+        "#8a96ab",
+        "#8996aa",
+        "#8895a9",
+        "#8794a9",
+        "#8693a8",
+        "#8592a8",
+        "#8492a7",
+        "#8391a6",
+        "#8290a6",
+        "#818fa5",
+        "#818ea5",
+        "#808ea4",
+        "#7f8da3",
+        "#7e8ca3",
+        "#7d8ba2",
+        "#7c8aa2",
+        "#7b8aa1",
+        "#7a89a1",
+        "#7988a0",
+        "#78879f",
+        "#77869f",
+        "#76869e",
+        "#76859e",
+        "#75849d",
+        "#74839d",
+        "#73829c",
+        "#72829b",
+        "#71819b",
+        "#70809a",
+        "#6f7f9a",
+        "#6e7f99",
+        "#6d7e99",
+        "#6c7d98",
+        "#6c7c98",
+        "#6b7c97",
+        "#6a7b97",
+        "#697a96",
+        "#687995",
+        "#677895",
+        "#667894",
+        "#657794",
+        "#647693",
+        "#637593",
+        "#637592",
+        "#627492",
+        "#617391",
+        "#607291",
+        "#5f7290",
+        "#5e7190",
+        "#5d708f",
+        "#5c6f8f",
+        "#5b6f8e",
+        "#5b6e8e",
+        "#5a6d8d",
+        "#596c8c",
+        "#586b8c",
+        "#576b8b",
+        "#566a8b",
+        "#55698a",
+        "#54688a",
+        "#536889",
+        "#536789",
+        "#526688",
+        "#516588",
+        "#506587",
+        "#4f6487",
+        "#4e6386",
+        "#4d6286",
+        "#4c6285",
+        "#4b6185",
+        "#4a6084",
+        "#4a5f84",
+        "#495f83",
+        "#485e83",
+        "#475d83",
+        "#465d82",
+        "#455c82",
+        "#445b81",
+        "#435a81",
+        "#425a80",
+        "#425980",
+        "#41587f",
+        "#40577f",
+        "#3f577e",
+        "#3e567e",
+        "#3d557d",
+        "#3c547d",
+        "#3b547c",
+        "#3a537c",
+        "#39527b",
+        "#39517b",
+        "#38517b",
+        "#37507a",
+        "#364f7a",
+        "#354e79",
+        "#344e79",
+        "#334d78",
+        "#324c78",
+        "#314b77",
+        "#304b77",
+        "#2f4a76",
+        "#2e4976",
+        "#2d4876",
+        "#2c4875",
+        "#2c4775",
+        "#2b4674",
+        "#2a4574",
+        "#294473",
+        "#284473",
+        "#274373",
+        "#264272",
+        "#254172",
+        "#244171",
+        "#234071",
+        "#223f70",
+        "#213e70",
+        "#203e70",
+        "#1f3d6f",
+        "#1e3c6f",
+        "#1d3b6e",
+        "#1c3a6e",
+        "#1b3a6e",
+        "#1a396d",
+        "#19386d",
+        "#17376c",
+        "#16366c",
+        "#15366c",
+        "#14356b",
+        "#13346b",
+        "#12336b",
+        "#10326a",
+        "#0f326a",
+        "#0e316a",
+        "#0d3069",
+        "#0b2f69",
+        "#0a2e68",
+        "#082d68",
+        "#072c68",
+        "#062c68",
+        "#042b67",
+        "#032a67",
+        "#022967",
+        "#012866",
+        "#002766",
+    ],
+    "flex_orange_seq": [
+        "#ffffff",
+        "#fefefe",
+        "#fefdfd",
+        "#fdfdfc",
+        "#fdfcfb",
+        "#fcfbfa",
+        "#fbfafa",
+        "#fbf9f9",
+        "#faf9f8",
+        "#faf8f7",
+        "#f9f7f6",
+        "#f8f6f5",
+        "#f8f6f4",
+        "#f7f5f3",
+        "#f7f4f2",
+        "#f6f3f1",
+        "#f5f2f1",
+        "#f5f2f0",
+        "#f4f1ef",
+        "#f3f0ee",
+        "#f3efed",
+        "#f2efec",
+        "#f2eeeb",
+        "#f1edea",
+        "#f1ece9",
+        "#f0ece8",
+        "#f0ebe7",
+        "#efeae6",
+        "#efe9e5",
+        "#eee9e4",
+        "#eee8e3",
+        "#ede7e2",
+        "#ede6e1",
+        "#ece5e0",
+        "#ece5df",
+        "#ebe4de",
+        "#ebe3dd",
+        "#eae2dc",
+        "#eae2db",
+        "#e9e1da",
+        "#e9e0d9",
+        "#e9dfd8",
+        "#e8dfd7",
+        "#e8ded6",
+        "#e7ddd5",
+        "#e7dcd4",
+        "#e6dbd3",
+        "#e6dbd2",
+        "#e6dad1",
+        "#e5d9d0",
+        "#e5d8cf",
+        "#e4d8ce",
+        "#e4d7cd",
+        "#e3d6cc",
+        "#e3d5cb",
+        "#e3d5ca",
+        "#e2d4c9",
+        "#e2d3c8",
+        "#e1d2c7",
+        "#e1d2c6",
+        "#e0d1c5",
+        "#e0d0c4",
+        "#e0cfc3",
+        "#dfcfc2",
+        "#dfcec1",
+        "#decdc0",
+        "#deccbf",
+        "#deccbe",
+        "#ddcbbd",
+        "#ddcabc",
+        "#dcc9bb",
+        "#dcc9ba",
+        "#dcc8b9",
+        "#dbc7b8",
+        "#dbc6b8",
+        "#dbc6b7",
+        "#dac5b6",
+        "#dac4b5",
+        "#d9c4b4",
+        "#d9c3b3",
+        "#d9c2b2",
+        "#d8c1b1",
+        "#d8c1b0",
+        "#d7c0af",
+        "#d7bfae",
+        "#d7bead",
+        "#d6beac",
+        "#d6bdab",
+        "#d6bcaa",
+        "#d5bba9",
+        "#d5bba8",
+        "#d4baa7",
+        "#d4b9a6",
+        "#d4b9a5",
+        "#d3b8a4",
+        "#d3b7a3",
+        "#d3b6a2",
+        "#d2b6a1",
+        "#d2b5a0",
+        "#d2b49f",
+        "#d1b49e",
+        "#d1b39d",
+        "#d0b29c",
+        "#d0b19b",
+        "#d0b19a",
+        "#cfb099",
+        "#cfaf99",
+        "#cfaf98",
+        "#ceae97",
+        "#cead96",
+        "#ceac95",
+        "#cdac94",
+        "#cdab93",
+        "#cdaa92",
+        "#ccaa91",
+        "#cca990",
+        "#cca88f",
+        "#cba78e",
+        "#cba78d",
+        "#caa68c",
+        "#caa58b",
+        "#caa58a",
+        "#c9a489",
+        "#c9a388",
+        "#c9a387",
+        "#c8a286",
+        "#c8a185",
+        "#c8a085",
+        "#c7a084",
+        "#c79f83",
+        "#c79e82",
+        "#c69e81",
+        "#c69d80",
+        "#c69c7f",
+        "#c59c7e",
+        "#c59b7d",
+        "#c59a7c",
+        "#c4997b",
+        "#c4997a",
+        "#c49879",
+        "#c39778",
+        "#c39777",
+        "#c29676",
+        "#c29575",
+        "#c29575",
+        "#c19474",
+        "#c19373",
+        "#c19372",
+        "#c09271",
+        "#c09170",
+        "#c0906f",
+        "#bf906e",
+        "#bf8f6d",
+        "#bf8e6c",
+        "#be8e6b",
+        "#be8d6a",
+        "#be8c69",
+        "#bd8c68",
+        "#bd8b67",
+        "#bd8a67",
+        "#bc8a66",
+        "#bc8965",
+        "#bc8864",
+        "#bb8863",
+        "#bb8762",
+        "#bb8661",
+        "#ba8660",
+        "#ba855f",
+        "#ba845e",
+        "#b9835d",
+        "#b9835c",
+        "#b8825b",
+        "#b8815b",
+        "#b8815a",
+        "#b78059",
+        "#b77f58",
+        "#b77f57",
+        "#b67e56",
+        "#b67d55",
+        "#b67d54",
+        "#b57c53",
+        "#b57b52",
+        "#b57b51",
+        "#b47a50",
+        "#b4794f",
+        "#b4794f",
+        "#b3784e",
+        "#b3774d",
+        "#b3774c",
+        "#b2764b",
+        "#b2754a",
+        "#b17549",
+        "#b17448",
+        "#b17347",
+        "#b07346",
+        "#b07245",
+        "#b07144",
+        "#af7144",
+        "#af7043",
+        "#af6f42",
+        "#ae6f41",
+        "#ae6e40",
+        "#ae6d3f",
+        "#ad6d3e",
+        "#ad6c3d",
+        "#ac6b3c",
+        "#ac6b3b",
+        "#ac6a3a",
+        "#ab6939",
+        "#ab6939",
+        "#ab6838",
+        "#aa6737",
+        "#aa6736",
+        "#aa6635",
+        "#a96534",
+        "#a96533",
+        "#a86432",
+        "#a86331",
+        "#a86330",
+        "#a7622f",
+        "#a7612e",
+        "#a7612d",
+        "#a6602c",
+        "#a65f2b",
+        "#a55f2a",
+        "#a55e2a",
+        "#a55d29",
+        "#a45d28",
+        "#a45c27",
+        "#a35b26",
+        "#a35b25",
+        "#a35a24",
+        "#a25923",
+        "#a25922",
+        "#a25821",
+        "#a15720",
+        "#a1571f",
+        "#a0561e",
+        "#a0551d",
+        "#a0551c",
+        "#9f541b",
+        "#9f531a",
+        "#9e5318",
+        "#9e5217",
+        "#9e5116",
+        "#9d5115",
+        "#9d5014",
+        "#9c4f13",
+        "#9c4f12",
+        "#9b4e10",
+        "#9b4d0f",
+        "#9b4d0e",
+        "#9a4c0c",
+        "#9a4b0b",
+        "#994b09",
+        "#994a08",
+    ],
+    "flex_red_seq": [
+        "#ffffff",
+        "#fefefe",
+        "#fefdfd",
+        "#fdfcfc",
+        "#fcfbfb",
+        "#fcfafa",
+        "#fbf9f9",
+        "#faf8f8",
+        "#faf7f7",
+        "#f9f6f6",
+        "#f8f5f5",
+        "#f8f4f5",
+        "#f7f3f4",
+        "#f6f2f3",
+        "#f5f2f2",
+        "#f5f1f1",
+        "#f4f0f0",
+        "#f3efef",
+        "#f3eeee",
+        "#f2eded",
+        "#f1ecec",
+        "#f1ebec",
+        "#f0eaeb",
+        "#efe9ea",
+        "#efe8e9",
+        "#eee7e8",
+        "#eee6e7",
+        "#ede5e6",
+        "#ece4e6",
+        "#ece3e5",
+        "#ebe2e4",
+        "#ebe1e3",
+        "#eae0e2",
+        "#eae0e1",
+        "#e9dfe0",
+        "#e9dedf",
+        "#e8dddf",
+        "#e8dcde",
+        "#e7dbdd",
+        "#e7dadc",
+        "#e6d9db",
+        "#e6d8da",
+        "#e5d7d9",
+        "#e5d6d8",
+        "#e4d5d7",
+        "#e4d4d7",
+        "#e3d3d6",
+        "#e3d2d5",
+        "#e2d1d4",
+        "#e2d0d3",
+        "#e1d0d2",
+        "#e1cfd1",
+        "#e0ced1",
+        "#e0cdd0",
+        "#dfcccf",
+        "#dfcbce",
+        "#decacd",
+        "#dec9cc",
+        "#ddc8cb",
+        "#ddc7cb",
+        "#dcc6ca",
+        "#dcc5c9",
+        "#dbc4c8",
+        "#dbc4c7",
+        "#dbc3c6",
+        "#dac2c5",
+        "#dac1c5",
+        "#d9c0c4",
+        "#d9bfc3",
+        "#d8bec2",
+        "#d8bdc1",
+        "#d7bcc0",
+        "#d7bbc0",
+        "#d7babf",
+        "#d6babe",
+        "#d6b9bd",
+        "#d5b8bc",
+        "#d5b7bb",
+        "#d4b6bb",
+        "#d4b5ba",
+        "#d4b4b9",
+        "#d3b3b8",
+        "#d3b2b7",
+        "#d2b1b6",
+        "#d2b0b6",
+        "#d1b0b5",
+        "#d1afb4",
+        "#d1aeb3",
+        "#d0adb2",
+        "#d0acb1",
+        "#cfabb1",
+        "#cfaab0",
+        "#cfa9af",
+        "#cea8ae",
+        "#cea8ad",
+        "#cda7ad",
+        "#cda6ac",
+        "#cca5ab",
+        "#cca4aa",
+        "#cca3a9",
+        "#cba2a9",
+        "#cba1a8",
+        "#caa0a7",
+        "#caa0a6",
+        "#ca9fa5",
+        "#c99ea5",
+        "#c99da4",
+        "#c89ca3",
+        "#c89ba2",
+        "#c89aa1",
+        "#c799a1",
+        "#c799a0",
+        "#c6989f",
+        "#c6979e",
+        "#c6969d",
+        "#c5959d",
+        "#c5949c",
+        "#c4939b",
+        "#c4929a",
+        "#c49299",
+        "#c39199",
+        "#c39098",
+        "#c28f97",
+        "#c28e96",
+        "#c28d96",
+        "#c18c95",
+        "#c18c94",
+        "#c18b93",
+        "#c08a92",
+        "#c08992",
+        "#bf8891",
+        "#bf8790",
+        "#bf868f",
+        "#be858f",
+        "#be858e",
+        "#bd848d",
+        "#bd838c",
+        "#bd828c",
+        "#bc818b",
+        "#bc808a",
+        "#bb7f89",
+        "#bb7f88",
+        "#bb7e88",
+        "#ba7d87",
+        "#ba7c86",
+        "#b97b85",
+        "#b97a85",
+        "#b97984",
+        "#b87983",
+        "#b87882",
+        "#b87782",
+        "#b77681",
+        "#b77580",
+        "#b6747f",
+        "#b6737f",
+        "#b6737e",
+        "#b5727d",
+        "#b5717c",
+        "#b4707c",
+        "#b46f7b",
+        "#b46e7a",
+        "#b36d79",
+        "#b36d79",
+        "#b26c78",
+        "#b26b77",
+        "#b26a76",
+        "#b16976",
+        "#b16875",
+        "#b06774",
+        "#b06773",
+        "#b06673",
+        "#af6572",
+        "#af6471",
+        "#ae6371",
+        "#ae6270",
+        "#ae616f",
+        "#ad616e",
+        "#ad606e",
+        "#ac5f6d",
+        "#ac5e6c",
+        "#ab5d6b",
+        "#ab5c6b",
+        "#ab5b6a",
+        "#aa5b69",
+        "#aa5a69",
+        "#a95968",
+        "#a95867",
+        "#a95766",
+        "#a85666",
+        "#a85565",
+        "#a75464",
+        "#a75463",
+        "#a65363",
+        "#a65262",
+        "#a65161",
+        "#a55061",
+        "#a54f60",
+        "#a44e5f",
+        "#a44d5e",
+        "#a34d5e",
+        "#a34c5d",
+        "#a34b5c",
+        "#a24a5c",
+        "#a2495b",
+        "#a1485a",
+        "#a1475a",
+        "#a04659",
+        "#a04558",
+        "#9f4557",
+        "#9f4457",
+        "#9f4356",
+        "#9e4255",
+        "#9e4155",
+        "#9d4054",
+        "#9d3f53",
+        "#9c3e52",
+        "#9c3d52",
+        "#9b3c51",
+        "#9b3b50",
+        "#9a3b50",
+        "#9a3a4f",
+        "#99394e",
+        "#99384e",
+        "#98374d",
+        "#98364c",
+        "#98354b",
+        "#97344b",
+        "#97334a",
+        "#963249",
+        "#963149",
+        "#953048",
+        "#952f47",
+        "#942e47",
+        "#942d46",
+        "#932c45",
+        "#932b45",
+        "#922a44",
+        "#922943",
+        "#912843",
+        "#912742",
+        "#902641",
+        "#902540",
+        "#8f2440",
+        "#8e223f",
+        "#8e213e",
+        "#8d203e",
+        "#8d1f3d",
+        "#8c1e3c",
+        "#8c1d3c",
+        "#8b1b3b",
+        "#8b1a3a",
+        "#8a193a",
+        "#8a1739",
+        "#891638",
+        "#891438",
+        "#881337",
+    ],
+    "flex_purple_seq": [
+        "#ffffff",
+        "#fefefe",
+        "#fdfdfd",
+        "#fcfcfd",
+        "#fbfbfc",
+        "#fafafb",
+        "#f9f9fa",
+        "#f8f8f9",
+        "#f7f7f9",
+        "#f6f6f8",
+        "#f5f5f7",
+        "#f4f4f6",
+        "#f3f3f6",
+        "#f2f2f5",
+        "#f1f1f4",
+        "#f0f0f3",
+        "#efeff3",
+        "#eeeef2",
+        "#ededf1",
+        "#ececf0",
+        "#ebebf0",
+        "#eaeaef",
+        "#e9e9ee",
+        "#e8e8ed",
+        "#e8e8ed",
+        "#e7e7ec",
+        "#e6e6eb",
+        "#e5e5eb",
+        "#e4e4ea",
+        "#e3e3e9",
+        "#e2e2e8",
+        "#e1e1e8",
+        "#e0e0e7",
+        "#dfdfe6",
+        "#dedee6",
+        "#dddde5",
+        "#dcdce4",
+        "#dbdbe4",
+        "#dadae3",
+        "#d9dae2",
+        "#d9d9e2",
+        "#d8d8e1",
+        "#d7d7e0",
+        "#d6d6e0",
+        "#d5d5df",
+        "#d4d4df",
+        "#d3d3de",
+        "#d2d2dd",
+        "#d1d1dd",
+        "#d0d1dc",
+        "#d0d0db",
+        "#cfcfdb",
+        "#ceceda",
+        "#cdcdd9",
+        "#ccccd9",
+        "#cbcbd8",
+        "#cacad8",
+        "#c9c9d7",
+        "#c8c9d6",
+        "#c8c8d6",
+        "#c7c7d5",
+        "#c6c6d5",
+        "#c5c5d4",
+        "#c4c4d3",
+        "#c3c3d3",
+        "#c2c2d2",
+        "#c2c2d2",
+        "#c1c1d1",
+        "#c0c0d1",
+        "#bfbfd0",
+        "#bebecf",
+        "#bdbdcf",
+        "#bcbcce",
+        "#bcbcce",
+        "#bbbbcd",
+        "#babacd",
+        "#b9b9cc",
+        "#b8b8cc",
+        "#b7b7cb",
+        "#b7b7ca",
+        "#b6b6ca",
+        "#b5b5c9",
+        "#b4b4c9",
+        "#b3b3c8",
+        "#b3b2c8",
+        "#b2b2c7",
+        "#b1b1c7",
+        "#b0b0c6",
+        "#afafc6",
+        "#aeaec5",
+        "#aeadc5",
+        "#adadc4",
+        "#acacc4",
+        "#ababc3",
+        "#aaaac3",
+        "#aaa9c2",
+        "#a9a8c2",
+        "#a8a8c1",
+        "#a7a7c1",
+        "#a6a6c0",
+        "#a6a5c0",
+        "#a5a4bf",
+        "#a4a4bf",
+        "#a3a3be",
+        "#a3a2be",
+        "#a2a1bd",
+        "#a1a0bd",
+        "#a0a0bc",
+        "#9f9fbc",
+        "#9f9ebb",
+        "#9e9dbb",
+        "#9d9cba",
+        "#9c9cba",
+        "#9c9bb9",
+        "#9b9ab9",
+        "#9a99b8",
+        "#9998b8",
+        "#9998b8",
+        "#9897b7",
+        "#9796b7",
+        "#9695b6",
+        "#9694b6",
+        "#9594b5",
+        "#9493b5",
+        "#9392b4",
+        "#9391b4",
+        "#9291b4",
+        "#9190b3",
+        "#908fb3",
+        "#908eb2",
+        "#8f8db2",
+        "#8e8db1",
+        "#8d8cb1",
+        "#8d8bb1",
+        "#8c8ab0",
+        "#8b8ab0",
+        "#8a89af",
+        "#8a88af",
+        "#8987ae",
+        "#8886ae",
+        "#8886ae",
+        "#8785ad",
+        "#8684ad",
+        "#8583ac",
+        "#8583ac",
+        "#8482ac",
+        "#8381ab",
+        "#8280ab",
+        "#8280ab",
+        "#817faa",
+        "#807eaa",
+        "#807da9",
+        "#7f7ca9",
+        "#7e7ca9",
+        "#7e7ba8",
+        "#7d7aa8",
+        "#7c79a8",
+        "#7b79a7",
+        "#7b78a7",
+        "#7a77a6",
+        "#7976a6",
+        "#7976a6",
+        "#7875a5",
+        "#7774a5",
+        "#7773a5",
+        "#7673a4",
+        "#7572a4",
+        "#7571a4",
+        "#7470a3",
+        "#736fa3",
+        "#736fa3",
+        "#726ea2",
+        "#716da2",
+        "#716ca2",
+        "#706ca1",
+        "#6f6ba1",
+        "#6f6aa1",
+        "#6e69a0",
+        "#6d69a0",
+        "#6d68a0",
+        "#6c679f",
+        "#6b669f",
+        "#6b669f",
+        "#6a659e",
+        "#69649e",
+        "#69639e",
+        "#68629d",
+        "#67629d",
+        "#67619d",
+        "#66609d",
+        "#655f9c",
+        "#655f9c",
+        "#645e9c",
+        "#635d9b",
+        "#635c9b",
+        "#625c9b",
+        "#625b9b",
+        "#615a9a",
+        "#60599a",
+        "#60589a",
+        "#5f589a",
+        "#5e5799",
+        "#5e5699",
+        "#5d5599",
+        "#5d5498",
+        "#5c5498",
+        "#5b5398",
+        "#5b5298",
+        "#5a5198",
+        "#595097",
+        "#595097",
+        "#584f97",
+        "#584e97",
+        "#574d96",
+        "#564c96",
+        "#564c96",
+        "#554b96",
+        "#554a95",
+        "#544995",
+        "#544895",
+        "#534795",
+        "#524795",
+        "#524694",
+        "#514594",
+        "#514494",
+        "#504394",
+        "#504294",
+        "#4f4294",
+        "#4e4193",
+        "#4e4093",
+        "#4d3f93",
+        "#4d3e93",
+        "#4c3d93",
+        "#4c3c93",
+        "#4b3b93",
+        "#4b3a92",
+        "#4a3992",
+        "#4a3892",
+        "#493892",
+        "#493792",
+        "#483692",
+        "#483592",
+        "#473492",
+        "#473391",
+        "#463291",
+        "#463191",
+        "#452f91",
+        "#452e91",
+        "#442d91",
+        "#442c91",
+        "#432b91",
+        "#432a91",
+        "#422991",
+        "#422891",
+        "#412691",
+        "#412591",
+    ],
+    "flex_grey_seq": [
+        "#ffffff",
+        "#fefefe",
+        "#fdfdfd",
+        "#fcfcfc",
+        "#fbfbfc",
+        "#fafafb",
+        "#f9f9fa",
+        "#f8f9f9",
+        "#f8f8f8",
+        "#f7f7f7",
+        "#f6f6f6",
+        "#f5f5f6",
+        "#f4f4f5",
+        "#f3f3f4",
+        "#f2f2f3",
+        "#f1f1f2",
+        "#f0f0f1",
+        "#eff0f1",
+        "#eeeff0",
+        "#eeeeef",
+        "#ededee",
+        "#ececed",
+        "#ebebec",
+        "#eaeaec",
+        "#e9e9eb",
+        "#e8e9ea",
+        "#e7e8e9",
+        "#e6e7e8",
+        "#e6e6e8",
+        "#e5e5e7",
+        "#e4e4e6",
+        "#e3e3e5",
+        "#e2e3e4",
+        "#e1e2e4",
+        "#e0e1e3",
+        "#dfe0e2",
+        "#dfdfe1",
+        "#dedee0",
+        "#dddde0",
+        "#dcdddf",
+        "#dbdcde",
+        "#dadbdd",
+        "#d9dadd",
+        "#d9d9dc",
+        "#d8d8db",
+        "#d7d8da",
+        "#d6d7da",
+        "#d5d6d9",
+        "#d4d5d8",
+        "#d4d4d7",
+        "#d3d4d6",
+        "#d2d3d6",
+        "#d1d2d5",
+        "#d0d1d4",
+        "#cfd0d3",
+        "#cfcfd3",
+        "#cecfd2",
+        "#cdced1",
+        "#cccdd0",
+        "#cbccd0",
+        "#cacbcf",
+        "#cacbce",
+        "#c9cace",
+        "#c8c9cd",
+        "#c7c8cc",
+        "#c6c8cb",
+        "#c6c7cb",
+        "#c5c6ca",
+        "#c4c5c9",
+        "#c3c4c8",
+        "#c2c4c8",
+        "#c2c3c7",
+        "#c1c2c6",
+        "#c0c1c6",
+        "#bfc0c5",
+        "#bec0c4",
+        "#bebfc3",
+        "#bdbec3",
+        "#bcbdc2",
+        "#bbbdc1",
+        "#babcc1",
+        "#babbc0",
+        "#b9babf",
+        "#b8babe",
+        "#b7b9be",
+        "#b7b8bd",
+        "#b6b7bc",
+        "#b5b7bc",
+        "#b4b6bb",
+        "#b3b5ba",
+        "#b3b4ba",
+        "#b2b4b9",
+        "#b1b3b8",
+        "#b0b2b8",
+        "#b0b1b7",
+        "#afb1b6",
+        "#aeb0b6",
+        "#adafb5",
+        "#adaeb4",
+        "#acaeb3",
+        "#abadb3",
+        "#aaacb2",
+        "#aaabb1",
+        "#a9abb1",
+        "#a8aab0",
+        "#a7a9af",
+        "#a7a8af",
+        "#a6a8ae",
+        "#a5a7ad",
+        "#a4a6ad",
+        "#a4a6ac",
+        "#a3a5ab",
+        "#a2a4ab",
+        "#a1a3aa",
+        "#a1a3aa",
+        "#a0a2a9",
+        "#9fa1a8",
+        "#9ea1a8",
+        "#9ea0a7",
+        "#9d9fa6",
+        "#9c9ea6",
+        "#9c9ea5",
+        "#9b9da4",
+        "#9a9ca4",
+        "#999ca3",
+        "#999ba2",
+        "#989aa2",
+        "#979aa1",
+        "#9799a0",
+        "#9698a0",
+        "#95979f",
+        "#94979f",
+        "#94969e",
+        "#93959d",
+        "#92959d",
+        "#92949c",
+        "#91939b",
+        "#90939b",
+        "#8f929a",
+        "#8f919a",
+        "#8e9199",
+        "#8d9098",
+        "#8d8f98",
+        "#8c8e97",
+        "#8b8e96",
+        "#8b8d96",
+        "#8a8c95",
+        "#898c95",
+        "#888b94",
+        "#888a93",
+        "#878a93",
+        "#868992",
+        "#868891",
+        "#858891",
+        "#848790",
+        "#848690",
+        "#83868f",
+        "#82858e",
+        "#82848e",
+        "#81848d",
+        "#80838d",
+        "#80828c",
+        "#7f828b",
+        "#7e818b",
+        "#7e808a",
+        "#7d808a",
+        "#7c7f89",
+        "#7b7e88",
+        "#7b7e88",
+        "#7a7d87",
+        "#797c87",
+        "#797c86",
+        "#787b85",
+        "#777a85",
+        "#777a84",
+        "#767984",
+        "#757983",
+        "#757882",
+        "#747782",
+        "#737781",
+        "#737681",
+        "#727580",
+        "#71757f",
+        "#71747f",
+        "#70737e",
+        "#70737e",
+        "#6f727d",
+        "#6e717d",
+        "#6e717c",
+        "#6d707b",
+        "#6c707b",
+        "#6c6f7a",
+        "#6b6e7a",
+        "#6a6e79",
+        "#6a6d79",
+        "#696c78",
+        "#686c77",
+        "#686b77",
+        "#676a76",
+        "#666a76",
+        "#666975",
+        "#656974",
+        "#646874",
+        "#646773",
+        "#636773",
+        "#636672",
+        "#626572",
+        "#616571",
+        "#616470",
+        "#606370",
+        "#5f636f",
+        "#5f626f",
+        "#5e626e",
+        "#5d616e",
+        "#5d606d",
+        "#5c606c",
+        "#5b5f6c",
+        "#5b5e6b",
+        "#5a5e6b",
+        "#5a5d6a",
+        "#595d6a",
+        "#585c69",
+        "#585b69",
+        "#575b68",
+        "#565a67",
+        "#565967",
+        "#555966",
+        "#555866",
+        "#545865",
+        "#535765",
+        "#535664",
+        "#525663",
+        "#515563",
+        "#515562",
+        "#505462",
+        "#4f5361",
+        "#4f5361",
+        "#4e5260",
+        "#4e5160",
+        "#4d515f",
+        "#4c505e",
+        "#4c505e",
+        "#4b4f5d",
+        "#4a4e5d",
+        "#4a4e5c",
+        "#494d5c",
+        "#494d5b",
+        "#484c5a",
+        "#474b5a",
+        "#474b59",
+        "#464a59",
+        "#454958",
+        "#454958",
+        "#444857",
+        "#444857",
+        "#434756",
+    ],
+}
+CATEGORICAL_PALETTES_HEX = {
+    "flex_distinct": [
+        "#176737",
+        "#FF7B0D",
+        "#979BAA",
+        "#F44E6A",
+        "#0062FF",
+        "#26AB5B",
+        "#6D3EF2",
+        "#F59E0B",
+    ]
+}
+DIVERGING_PALETTES_HEX = {
+    "flex_BuRd": [
+        "#002766",
+        "#022967",
+        "#052b67",
+        "#072d68",
+        "#0a2e69",
+        "#0d3069",
+        "#10326a",
+        "#12346b",
+        "#15356c",
+        "#17376c",
+        "#1a396d",
+        "#1c3a6e",
+        "#1e3c6f",
+        "#203e70",
+        "#223f71",
+        "#244171",
+        "#264372",
+        "#284473",
+        "#2a4674",
+        "#2c4775",
+        "#2e4976",
+        "#304a77",
+        "#324c78",
+        "#344e79",
+        "#364f7a",
+        "#38517b",
+        "#3a527c",
+        "#3c547d",
+        "#3e557e",
+        "#3f577f",
+        "#415980",
+        "#435a80",
+        "#455c81",
+        "#475d83",
+        "#495f84",
+        "#4b6085",
+        "#4c6286",
+        "#4e6387",
+        "#506588",
+        "#526789",
+        "#54688a",
+        "#566a8b",
+        "#586b8c",
+        "#5a6d8d",
+        "#5b6e8e",
+        "#5d708f",
+        "#5f7290",
+        "#617391",
+        "#637592",
+        "#657694",
+        "#677895",
+        "#687a96",
+        "#6a7b97",
+        "#6c7d98",
+        "#6e7e99",
+        "#70809a",
+        "#72829b",
+        "#74839d",
+        "#76859e",
+        "#78879f",
+        "#7a88a0",
+        "#7b8aa1",
+        "#7d8ca3",
+        "#7f8da4",
+        "#818fa5",
+        "#8391a6",
+        "#8592a8",
+        "#8794a9",
+        "#8996aa",
+        "#8b97ab",
+        "#8d99ad",
+        "#8f9bae",
+        "#919daf",
+        "#939eb1",
+        "#95a0b2",
+        "#97a2b3",
+        "#99a4b4",
+        "#9ba5b6",
+        "#9da7b7",
+        "#9fa9b9",
+        "#a1abba",
+        "#a3acbb",
+        "#a5aebd",
+        "#a7b0be",
+        "#a9b2c0",
+        "#abb4c1",
+        "#adb6c2",
+        "#afb7c4",
+        "#b1b9c5",
+        "#b4bbc7",
+        "#b6bdc8",
+        "#b8bfca",
+        "#bac1cb",
+        "#bcc3cd",
+        "#bec5ce",
+        "#c0c6d0",
+        "#c3c8d1",
+        "#c5cad3",
+        "#c7ccd5",
+        "#c9ced6",
+        "#cbd0d8",
+        "#ced2d9",
+        "#d0d4db",
+        "#d2d6dd",
+        "#d4d8de",
+        "#d7dae0",
+        "#d9dce2",
+        "#dbdee3",
+        "#dee0e5",
+        "#e0e3e7",
+        "#e2e5e9",
+        "#e4e7ea",
+        "#e7e9ec",
+        "#e9ebee",
+        "#ecedf0",
+        "#eeeff2",
+        "#f0f2f3",
+        "#f3f4f5",
+        "#f5f6f7",
+        "#f8f8f9",
+        "#fafafb",
+        "#fdfdfd",
+        "#FFFFFF",
+        "#fefdfd",
+        "#fcfbfb",
+        "#fbf9f9",
+        "#f9f7f7",
+        "#f8f5f5",
+        "#f6f3f3",
+        "#f5f1f1",
+        "#f4efef",
+        "#f2edee",
+        "#f1ebec",
+        "#efe9ea",
+        "#eee7e8",
+        "#ede5e6",
+        "#ece3e4",
+        "#ebe1e3",
+        "#e9dfe1",
+        "#e8dddf",
+        "#e7dbdd",
+        "#e6d9db",
+        "#e5d7d9",
+        "#e4d5d8",
+        "#e3d3d6",
+        "#e2d1d4",
+        "#e1cfd2",
+        "#e0cdd0",
+        "#dfcccf",
+        "#decacd",
+        "#ddc8cb",
+        "#dcc6c9",
+        "#dbc4c7",
+        "#dac2c6",
+        "#d9c0c4",
+        "#d8bec2",
+        "#d7bcc0",
+        "#d6babf",
+        "#d6b8bd",
+        "#d5b6bb",
+        "#d4b5b9",
+        "#d3b3b8",
+        "#d2b1b6",
+        "#d1afb4",
+        "#d0adb2",
+        "#cfabb1",
+        "#cfa9af",
+        "#cea8ad",
+        "#cda6ac",
+        "#cca4aa",
+        "#cba2a8",
+        "#caa0a7",
+        "#c99ea5",
+        "#c99ca3",
+        "#c89ba2",
+        "#c799a0",
+        "#c6979e",
+        "#c5959d",
+        "#c4939b",
+        "#c49199",
+        "#c39098",
+        "#c28e96",
+        "#c18c94",
+        "#c08a93",
+        "#c08891",
+        "#bf8790",
+        "#be858e",
+        "#bd838c",
+        "#bc818b",
+        "#bb7f89",
+        "#bb7e88",
+        "#ba7c86",
+        "#b97a84",
+        "#b87883",
+        "#b77681",
+        "#b67580",
+        "#b6737e",
+        "#b5717d",
+        "#b46f7b",
+        "#b36d79",
+        "#b26c78",
+        "#b26a76",
+        "#b16875",
+        "#b06673",
+        "#af6572",
+        "#ae6370",
+        "#ad616f",
+        "#ac5f6d",
+        "#ac5d6c",
+        "#ab5c6a",
+        "#aa5a69",
+        "#a95867",
+        "#a85666",
+        "#a75464",
+        "#a65263",
+        "#a55161",
+        "#a44f60",
+        "#a44d5e",
+        "#a34b5d",
+        "#a2495b",
+        "#a1475a",
+        "#a04658",
+        "#9f4457",
+        "#9e4255",
+        "#9d4054",
+        "#9c3e52",
+        "#9b3c51",
+        "#9a3a4f",
+        "#99384e",
+        "#98364c",
+        "#97344b",
+        "#96324a",
+        "#953048",
+        "#942e47",
+        "#932c45",
+        "#922a44",
+        "#912842",
+        "#902541",
+        "#8f233f",
+        "#8e213e",
+        "#8d1e3d",
+        "#8b1c3b",
+        "#8a193a",
+        "#891638",
+        "#881337",
+    ],
+    "flex_RdBu": [
+        "#881337",
+        "#891638",
+        "#8a193a",
+        "#8b1c3b",
+        "#8d1e3d",
+        "#8e213e",
+        "#8f233f",
+        "#902541",
+        "#912842",
+        "#922a44",
+        "#932c45",
+        "#942e47",
+        "#953048",
+        "#96324a",
+        "#97344b",
+        "#98364c",
+        "#99384e",
+        "#9a3a4f",
+        "#9b3c51",
+        "#9c3e52",
+        "#9d4054",
+        "#9e4255",
+        "#9f4457",
+        "#a04658",
+        "#a1475a",
+        "#a2495b",
+        "#a34b5d",
+        "#a44d5e",
+        "#a44f60",
+        "#a55161",
+        "#a65263",
+        "#a75464",
+        "#a85666",
+        "#a95867",
+        "#aa5a69",
+        "#ab5c6a",
+        "#ac5d6c",
+        "#ac5f6d",
+        "#ad616f",
+        "#ae6370",
+        "#af6572",
+        "#b06673",
+        "#b16875",
+        "#b26a76",
+        "#b26c78",
+        "#b36d79",
+        "#b46f7b",
+        "#b5717d",
+        "#b6737e",
+        "#b67580",
+        "#b77681",
+        "#b87883",
+        "#b97a84",
+        "#ba7c86",
+        "#bb7e88",
+        "#bb7f89",
+        "#bc818b",
+        "#bd838c",
+        "#be858e",
+        "#bf8790",
+        "#c08891",
+        "#c08a93",
+        "#c18c94",
+        "#c28e96",
+        "#c39098",
+        "#c49199",
+        "#c4939b",
+        "#c5959d",
+        "#c6979e",
+        "#c799a0",
+        "#c89ba2",
+        "#c99ca3",
+        "#c99ea5",
+        "#caa0a7",
+        "#cba2a8",
+        "#cca4aa",
+        "#cda6ac",
+        "#cea8ad",
+        "#cfa9af",
+        "#cfabb1",
+        "#d0adb2",
+        "#d1afb4",
+        "#d2b1b6",
+        "#d3b3b8",
+        "#d4b5b9",
+        "#d5b6bb",
+        "#d6b8bd",
+        "#d6babf",
+        "#d7bcc0",
+        "#d8bec2",
+        "#d9c0c4",
+        "#dac2c6",
+        "#dbc4c7",
+        "#dcc6c9",
+        "#ddc8cb",
+        "#decacd",
+        "#dfcccf",
+        "#e0cdd0",
+        "#e1cfd2",
+        "#e2d1d4",
+        "#e3d3d6",
+        "#e4d5d8",
+        "#e5d7d9",
+        "#e6d9db",
+        "#e7dbdd",
+        "#e8dddf",
+        "#e9dfe1",
+        "#ebe1e3",
+        "#ece3e4",
+        "#ede5e6",
+        "#eee7e8",
+        "#efe9ea",
+        "#f1ebec",
+        "#f2edee",
+        "#f4efef",
+        "#f5f1f1",
+        "#f6f3f3",
+        "#f8f5f5",
+        "#f9f7f7",
+        "#fbf9f9",
+        "#fcfbfb",
+        "#fefdfd",
+        "#FFFFFF",
+        "#fdfdfd",
+        "#fafafb",
+        "#f8f8f9",
+        "#f5f6f7",
+        "#f3f4f5",
+        "#f0f2f3",
+        "#eeeff2",
+        "#ecedf0",
+        "#e9ebee",
+        "#e7e9ec",
+        "#e4e7ea",
+        "#e2e5e9",
+        "#e0e3e7",
+        "#dee0e5",
+        "#dbdee3",
+        "#d9dce2",
+        "#d7dae0",
+        "#d4d8de",
+        "#d2d6dd",
+        "#d0d4db",
+        "#ced2d9",
+        "#cbd0d8",
+        "#c9ced6",
+        "#c7ccd5",
+        "#c5cad3",
+        "#c3c8d1",
+        "#c0c6d0",
+        "#bec5ce",
+        "#bcc3cd",
+        "#bac1cb",
+        "#b8bfca",
+        "#b6bdc8",
+        "#b4bbc7",
+        "#b1b9c5",
+        "#afb7c4",
+        "#adb6c2",
+        "#abb4c1",
+        "#a9b2c0",
+        "#a7b0be",
+        "#a5aebd",
+        "#a3acbb",
+        "#a1abba",
+        "#9fa9b9",
+        "#9da7b7",
+        "#9ba5b6",
+        "#99a4b4",
+        "#97a2b3",
+        "#95a0b2",
+        "#939eb1",
+        "#919daf",
+        "#8f9bae",
+        "#8d99ad",
+        "#8b97ab",
+        "#8996aa",
+        "#8794a9",
+        "#8592a8",
+        "#8391a6",
+        "#818fa5",
+        "#7f8da4",
+        "#7d8ca3",
+        "#7b8aa1",
+        "#7a88a0",
+        "#78879f",
+        "#76859e",
+        "#74839d",
+        "#72829b",
+        "#70809a",
+        "#6e7e99",
+        "#6c7d98",
+        "#6a7b97",
+        "#687a96",
+        "#677895",
+        "#657694",
+        "#637592",
+        "#617391",
+        "#5f7290",
+        "#5d708f",
+        "#5b6e8e",
+        "#5a6d8d",
+        "#586b8c",
+        "#566a8b",
+        "#54688a",
+        "#526789",
+        "#506588",
+        "#4e6387",
+        "#4c6286",
+        "#4b6085",
+        "#495f84",
+        "#475d83",
+        "#455c81",
+        "#435a80",
+        "#415980",
+        "#3f577f",
+        "#3e557e",
+        "#3c547d",
+        "#3a527c",
+        "#38517b",
+        "#364f7a",
+        "#344e79",
+        "#324c78",
+        "#304a77",
+        "#2e4976",
+        "#2c4775",
+        "#2a4674",
+        "#284473",
+        "#264372",
+        "#244171",
+        "#223f71",
+        "#203e70",
+        "#1e3c6f",
+        "#1c3a6e",
+        "#1a396d",
+        "#17376c",
+        "#15356c",
+        "#12346b",
+        "#10326a",
+        "#0d3069",
+        "#0a2e69",
+        "#072d68",
+        "#052b67",
+        "#022967",
+        "#002766",
+    ],
+    "flex_GrPu": [
+        "#0f4424",
+        "#124526",
+        "#144727",
+        "#174829",
+        "#19492a",
+        "#1b4b2c",
+        "#1d4c2d",
+        "#1f4d2f",
+        "#214f30",
+        "#235032",
+        "#255234",
+        "#275335",
+        "#295437",
+        "#2b5638",
+        "#2d573a",
+        "#2f583b",
+        "#315a3d",
+        "#335b3e",
+        "#355c40",
+        "#365e42",
+        "#385f43",
+        "#3a6045",
+        "#3c6246",
+        "#3e6348",
+        "#3f644a",
+        "#41664b",
+        "#43674d",
+        "#45684e",
+        "#476a50",
+        "#486b52",
+        "#4a6c53",
+        "#4c6e55",
+        "#4e6f56",
+        "#4f7058",
+        "#51725a",
+        "#53735b",
+        "#55745d",
+        "#57765f",
+        "#587760",
+        "#5a7962",
+        "#5c7a63",
+        "#5e7b65",
+        "#5f7d67",
+        "#617e68",
+        "#637f6a",
+        "#65816c",
+        "#67826d",
+        "#68846f",
+        "#6a8571",
+        "#6c8672",
+        "#6e8874",
+        "#6f8976",
+        "#718b78",
+        "#738c79",
+        "#758e7b",
+        "#778f7d",
+        "#79907e",
+        "#7a9280",
+        "#7c9382",
+        "#7e9584",
+        "#809685",
+        "#829887",
+        "#849989",
+        "#859b8b",
+        "#879c8c",
+        "#899d8e",
+        "#8b9f90",
+        "#8da092",
+        "#8fa294",
+        "#91a395",
+        "#92a597",
+        "#94a699",
+        "#96a89b",
+        "#98aa9d",
+        "#9aab9e",
+        "#9cada0",
+        "#9eaea2",
+        "#a0b0a4",
+        "#a2b1a6",
+        "#a4b3a8",
+        "#a6b4aa",
+        "#a8b6ab",
+        "#aab8ad",
+        "#acb9af",
+        "#aebbb1",
+        "#b0bcb3",
+        "#b2beb5",
+        "#b4c0b7",
+        "#b6c1b9",
+        "#b8c3bb",
+        "#bac5bd",
+        "#bcc6bf",
+        "#bec8c1",
+        "#c0cac2",
+        "#c2cbc4",
+        "#c4cdc6",
+        "#c6cfc8",
+        "#c8d0ca",
+        "#cad2cc",
+        "#ccd4ce",
+        "#ced6d0",
+        "#d0d7d2",
+        "#d3d9d5",
+        "#d5dbd7",
+        "#d7ddd9",
+        "#d9dfdb",
+        "#dbe0dd",
+        "#dde2df",
+        "#dfe4e1",
+        "#e2e6e3",
+        "#e4e8e5",
+        "#e6eae7",
+        "#e8ebe9",
+        "#ebedeb",
+        "#edefee",
+        "#eff1f0",
+        "#f1f3f2",
+        "#f4f5f4",
+        "#f6f7f6",
+        "#f8f9f8",
+        "#fafbfb",
+        "#fdfdfd",
+        "#FFFFFF",
+        "#fdfdfd",
+        "#fbfbfc",
+        "#f9f9fa",
+        "#f7f7f8",
+        "#f5f5f7",
+        "#f3f3f5",
+        "#f1f0f4",
+        "#efeef2",
+        "#ececf0",
+        "#eaeaef",
+        "#e8e8ed",
+        "#e6e6ec",
+        "#e4e5ea",
+        "#e3e3e9",
+        "#e1e1e8",
+        "#dfdfe6",
+        "#dddde5",
+        "#dbdbe3",
+        "#d9d9e2",
+        "#d7d7e1",
+        "#d5d5df",
+        "#d3d3de",
+        "#d1d1dd",
+        "#cfd0db",
+        "#ceceda",
+        "#ccccd9",
+        "#cacad7",
+        "#c8c8d6",
+        "#c6c6d5",
+        "#c4c4d4",
+        "#c3c3d2",
+        "#c1c1d1",
+        "#bfbfd0",
+        "#bdbdcf",
+        "#bcbcce",
+        "#babacd",
+        "#b8b8cb",
+        "#b6b6ca",
+        "#b5b4c9",
+        "#b3b3c8",
+        "#b1b1c7",
+        "#afafc6",
+        "#aeaec5",
+        "#acacc4",
+        "#aaaac3",
+        "#a9a8c1",
+        "#a7a7c0",
+        "#a5a5bf",
+        "#a4a3be",
+        "#a2a2bd",
+        "#a1a0bc",
+        "#9f9ebb",
+        "#9d9dba",
+        "#9c9bb9",
+        "#9a99b8",
+        "#9898b8",
+        "#9796b7",
+        "#9594b6",
+        "#9493b5",
+        "#9291b4",
+        "#918fb3",
+        "#8f8eb2",
+        "#8e8cb1",
+        "#8c8ab0",
+        "#8b89af",
+        "#8987af",
+        "#8786ae",
+        "#8684ad",
+        "#8482ac",
+        "#8381ab",
+        "#817faa",
+        "#807eaa",
+        "#7f7ca9",
+        "#7d7aa8",
+        "#7c79a7",
+        "#7a77a6",
+        "#7976a6",
+        "#7774a5",
+        "#7672a4",
+        "#7471a4",
+        "#736fa3",
+        "#726ea2",
+        "#706ca1",
+        "#6f6aa1",
+        "#6d69a0",
+        "#6c679f",
+        "#6b669f",
+        "#69649e",
+        "#68629d",
+        "#67619d",
+        "#655f9c",
+        "#645e9c",
+        "#635c9b",
+        "#615a9a",
+        "#60599a",
+        "#5f5799",
+        "#5d5599",
+        "#5c5498",
+        "#5b5298",
+        "#595097",
+        "#584f97",
+        "#574d96",
+        "#564b96",
+        "#554a95",
+        "#534895",
+        "#524695",
+        "#514494",
+        "#504394",
+        "#4f4193",
+        "#4d3f93",
+        "#4c3d93",
+        "#4b3b92",
+        "#4a3992",
+        "#493792",
+        "#483592",
+        "#473392",
+        "#463191",
+        "#452f91",
+        "#442d91",
+        "#432a91",
+        "#422891",
+        "#412591",
+    ],
+    "flex_PuGr": [
+        "#412591",
+        "#422891",
+        "#432a91",
+        "#442d91",
+        "#452f91",
+        "#463191",
+        "#473392",
+        "#483592",
+        "#493792",
+        "#4a3992",
+        "#4b3b92",
+        "#4c3d93",
+        "#4d3f93",
+        "#4f4193",
+        "#504394",
+        "#514494",
+        "#524695",
+        "#534895",
+        "#554a95",
+        "#564b96",
+        "#574d96",
+        "#584f97",
+        "#595097",
+        "#5b5298",
+        "#5c5498",
+        "#5d5599",
+        "#5f5799",
+        "#60599a",
+        "#615a9a",
+        "#635c9b",
+        "#645e9c",
+        "#655f9c",
+        "#67619d",
+        "#68629d",
+        "#69649e",
+        "#6b669f",
+        "#6c679f",
+        "#6d69a0",
+        "#6f6aa1",
+        "#706ca1",
+        "#726ea2",
+        "#736fa3",
+        "#7471a4",
+        "#7672a4",
+        "#7774a5",
+        "#7976a6",
+        "#7a77a6",
+        "#7c79a7",
+        "#7d7aa8",
+        "#7f7ca9",
+        "#807eaa",
+        "#817faa",
+        "#8381ab",
+        "#8482ac",
+        "#8684ad",
+        "#8786ae",
+        "#8987af",
+        "#8b89af",
+        "#8c8ab0",
+        "#8e8cb1",
+        "#8f8eb2",
+        "#918fb3",
+        "#9291b4",
+        "#9493b5",
+        "#9594b6",
+        "#9796b7",
+        "#9898b8",
+        "#9a99b8",
+        "#9c9bb9",
+        "#9d9dba",
+        "#9f9ebb",
+        "#a1a0bc",
+        "#a2a2bd",
+        "#a4a3be",
+        "#a5a5bf",
+        "#a7a7c0",
+        "#a9a8c1",
+        "#aaaac3",
+        "#acacc4",
+        "#aeaec5",
+        "#afafc6",
+        "#b1b1c7",
+        "#b3b3c8",
+        "#b5b4c9",
+        "#b6b6ca",
+        "#b8b8cb",
+        "#babacd",
+        "#bcbcce",
+        "#bdbdcf",
+        "#bfbfd0",
+        "#c1c1d1",
+        "#c3c3d2",
+        "#c4c4d4",
+        "#c6c6d5",
+        "#c8c8d6",
+        "#cacad7",
+        "#ccccd9",
+        "#ceceda",
+        "#cfd0db",
+        "#d1d1dd",
+        "#d3d3de",
+        "#d5d5df",
+        "#d7d7e1",
+        "#d9d9e2",
+        "#dbdbe3",
+        "#dddde5",
+        "#dfdfe6",
+        "#e1e1e8",
+        "#e3e3e9",
+        "#e4e5ea",
+        "#e6e6ec",
+        "#e8e8ed",
+        "#eaeaef",
+        "#ececf0",
+        "#efeef2",
+        "#f1f0f4",
+        "#f3f3f5",
+        "#f5f5f7",
+        "#f7f7f8",
+        "#f9f9fa",
+        "#fbfbfc",
+        "#fdfdfd",
+        "#FFFFFF",
+        "#fdfdfd",
+        "#fafbfb",
+        "#f8f9f8",
+        "#f6f7f6",
+        "#f4f5f4",
+        "#f1f3f2",
+        "#eff1f0",
+        "#edefee",
+        "#ebedeb",
+        "#e8ebe9",
+        "#e6eae7",
+        "#e4e8e5",
+        "#e2e6e3",
+        "#dfe4e1",
+        "#dde2df",
+        "#dbe0dd",
+        "#d9dfdb",
+        "#d7ddd9",
+        "#d5dbd7",
+        "#d3d9d5",
+        "#d0d7d2",
+        "#ced6d0",
+        "#ccd4ce",
+        "#cad2cc",
+        "#c8d0ca",
+        "#c6cfc8",
+        "#c4cdc6",
+        "#c2cbc4",
+        "#c0cac2",
+        "#bec8c1",
+        "#bcc6bf",
+        "#bac5bd",
+        "#b8c3bb",
+        "#b6c1b9",
+        "#b4c0b7",
+        "#b2beb5",
+        "#b0bcb3",
+        "#aebbb1",
+        "#acb9af",
+        "#aab8ad",
+        "#a8b6ab",
+        "#a6b4aa",
+        "#a4b3a8",
+        "#a2b1a6",
+        "#a0b0a4",
+        "#9eaea2",
+        "#9cada0",
+        "#9aab9e",
+        "#98aa9d",
+        "#96a89b",
+        "#94a699",
+        "#92a597",
+        "#91a395",
+        "#8fa294",
+        "#8da092",
+        "#8b9f90",
+        "#899d8e",
+        "#879c8c",
+        "#859b8b",
+        "#849989",
+        "#829887",
+        "#809685",
+        "#7e9584",
+        "#7c9382",
+        "#7a9280",
+        "#79907e",
+        "#778f7d",
+        "#758e7b",
+        "#738c79",
+        "#718b78",
+        "#6f8976",
+        "#6e8874",
+        "#6c8672",
+        "#6a8571",
+        "#68846f",
+        "#67826d",
+        "#65816c",
+        "#637f6a",
+        "#617e68",
+        "#5f7d67",
+        "#5e7b65",
+        "#5c7a63",
+        "#5a7962",
+        "#587760",
+        "#57765f",
+        "#55745d",
+        "#53735b",
+        "#51725a",
+        "#4f7058",
+        "#4e6f56",
+        "#4c6e55",
+        "#4a6c53",
+        "#486b52",
+        "#476a50",
+        "#45684e",
+        "#43674d",
+        "#41664b",
+        "#3f644a",
+        "#3e6348",
+        "#3c6246",
+        "#3a6045",
+        "#385f43",
+        "#365e42",
+        "#355c40",
+        "#335b3e",
+        "#315a3d",
+        "#2f583b",
+        "#2d573a",
+        "#2b5638",
+        "#295437",
+        "#275335",
+        "#255234",
+        "#235032",
+        "#214f30",
+        "#1f4d2f",
+        "#1d4c2d",
+        "#1b4b2c",
+        "#19492a",
+        "#174829",
+        "#144727",
+        "#124526",
+        "#0f4424",
+    ],
+    "flex_TuOr": [
+        "#134e4a",
+        "#164f4b",
+        "#19504d",
+        "#1b524e",
+        "#1e534f",
+        "#205450",
+        "#225552",
+        "#255753",
+        "#275854",
+        "#295955",
+        "#2b5a57",
+        "#2d5c58",
+        "#2f5d59",
+        "#315e5a",
+        "#335f5c",
+        "#35615d",
+        "#37625e",
+        "#39635f",
+        "#3b6461",
+        "#3c6662",
+        "#3e6763",
+        "#406865",
+        "#426966",
+        "#446b67",
+        "#456c68",
+        "#476d6a",
+        "#496e6b",
+        "#4b706c",
+        "#4d716e",
+        "#4e726f",
+        "#507370",
+        "#527572",
+        "#547673",
+        "#557774",
+        "#577975",
+        "#597a77",
+        "#5b7b78",
+        "#5d7c79",
+        "#5e7e7b",
+        "#607f7c",
+        "#62807d",
+        "#64827f",
+        "#658380",
+        "#678482",
+        "#698683",
+        "#6b8784",
+        "#6c8886",
+        "#6e8a87",
+        "#708b88",
+        "#728c8a",
+        "#738e8b",
+        "#758f8c",
+        "#77908e",
+        "#79928f",
+        "#7a9391",
+        "#7c9492",
+        "#7e9693",
+        "#809795",
+        "#819896",
+        "#839a98",
+        "#859b99",
+        "#879d9b",
+        "#899e9c",
+        "#8a9f9d",
+        "#8ca19f",
+        "#8ea2a0",
+        "#90a4a2",
+        "#92a5a3",
+        "#93a7a5",
+        "#95a8a6",
+        "#97a9a8",
+        "#99aba9",
+        "#9bacab",
+        "#9daeac",
+        "#9eafae",
+        "#a0b1af",
+        "#a2b2b1",
+        "#a4b4b2",
+        "#a6b5b4",
+        "#a8b7b5",
+        "#aab8b7",
+        "#acbab8",
+        "#adbbba",
+        "#afbdbb",
+        "#b1bebd",
+        "#b3c0bf",
+        "#b5c1c0",
+        "#b7c3c2",
+        "#b9c5c3",
+        "#bbc6c5",
+        "#bdc8c7",
+        "#bfc9c8",
+        "#c1cbca",
+        "#c3cccc",
+        "#c5cecd",
+        "#c7d0cf",
+        "#c9d1d1",
+        "#cbd3d2",
+        "#cdd5d4",
+        "#cfd6d6",
+        "#d1d8d7",
+        "#d3dad9",
+        "#d5dbdb",
+        "#d7dddc",
+        "#d9dfde",
+        "#dbe0e0",
+        "#dde2e2",
+        "#dfe4e3",
+        "#e1e6e5",
+        "#e3e7e7",
+        "#e5e9e9",
+        "#e7ebeb",
+        "#e9edec",
+        "#ebeeee",
+        "#eef0f0",
+        "#f0f2f2",
+        "#f2f4f4",
+        "#f4f6f6",
+        "#f6f8f7",
+        "#f8f9f9",
+        "#fbfbfb",
+        "#fdfdfd",
+        "#FFFFFF",
+        "#fefdfd",
+        "#fcfcfb",
+        "#fbfaf9",
+        "#faf8f7",
+        "#f9f7f6",
+        "#f7f5f4",
+        "#f6f4f2",
+        "#f5f2f0",
+        "#f4f0ee",
+        "#f2efec",
+        "#f1edea",
+        "#f0ece8",
+        "#efeae6",
+        "#eee8e4",
+        "#ede7e2",
+        "#ece5df",
+        "#ebe4dd",
+        "#eae2db",
+        "#e9e0d9",
+        "#e8dfd7",
+        "#e7ddd5",
+        "#e6dcd3",
+        "#e5dad1",
+        "#e5d8cf",
+        "#e4d7cd",
+        "#e3d5cb",
+        "#e2d4c9",
+        "#e1d2c7",
+        "#e0d0c5",
+        "#dfcfc3",
+        "#dfcdc1",
+        "#deccbe",
+        "#ddcabc",
+        "#dcc9ba",
+        "#dbc7b8",
+        "#dac6b6",
+        "#dac4b4",
+        "#d9c2b2",
+        "#d8c1b0",
+        "#d7bfae",
+        "#d6beac",
+        "#d6bcaa",
+        "#d5bba8",
+        "#d4b9a6",
+        "#d3b8a4",
+        "#d3b6a2",
+        "#d2b5a0",
+        "#d1b39e",
+        "#d0b29c",
+        "#d0b09a",
+        "#cfaf98",
+        "#cead96",
+        "#cdac94",
+        "#cdaa92",
+        "#cca990",
+        "#cba78e",
+        "#caa68c",
+        "#caa48a",
+        "#c9a388",
+        "#c8a286",
+        "#c7a084",
+        "#c79f82",
+        "#c69d80",
+        "#c59c7e",
+        "#c59a7c",
+        "#c4997a",
+        "#c39778",
+        "#c29676",
+        "#c29474",
+        "#c19372",
+        "#c09270",
+        "#c0906e",
+        "#bf8f6c",
+        "#be8d6b",
+        "#bd8c69",
+        "#bd8a67",
+        "#bc8965",
+        "#bb8863",
+        "#bb8661",
+        "#ba855f",
+        "#b9835d",
+        "#b8825b",
+        "#b88059",
+        "#b77f57",
+        "#b67e55",
+        "#b57c53",
+        "#b57b51",
+        "#b47950",
+        "#b3784e",
+        "#b3774c",
+        "#b2754a",
+        "#b17448",
+        "#b07246",
+        "#b07144",
+        "#af7042",
+        "#ae6e40",
+        "#ad6d3e",
+        "#ad6b3c",
+        "#ac6a3a",
+        "#ab6938",
+        "#aa6737",
+        "#a96635",
+        "#a96433",
+        "#a86331",
+        "#a7622f",
+        "#a6602d",
+        "#a65f2b",
+        "#a55d29",
+        "#a45c27",
+        "#a35b25",
+        "#a25923",
+        "#a15821",
+        "#a1571f",
+        "#a0551c",
+        "#9f541a",
+        "#9e5218",
+        "#9d5116",
+        "#9c5013",
+        "#9c4e11",
+        "#9b4d0e",
+        "#9a4b0b",
+        "#994a08",
+    ],
+    "flex_OrTu": [
+        "#994a08",
+        "#9a4b0b",
+        "#9b4d0e",
+        "#9c4e11",
+        "#9c5013",
+        "#9d5116",
+        "#9e5218",
+        "#9f541a",
+        "#a0551c",
+        "#a1571f",
+        "#a15821",
+        "#a25923",
+        "#a35b25",
+        "#a45c27",
+        "#a55d29",
+        "#a65f2b",
+        "#a6602d",
+        "#a7622f",
+        "#a86331",
+        "#a96433",
+        "#a96635",
+        "#aa6737",
+        "#ab6938",
+        "#ac6a3a",
+        "#ad6b3c",
+        "#ad6d3e",
+        "#ae6e40",
+        "#af7042",
+        "#b07144",
+        "#b07246",
+        "#b17448",
+        "#b2754a",
+        "#b3774c",
+        "#b3784e",
+        "#b47950",
+        "#b57b51",
+        "#b57c53",
+        "#b67e55",
+        "#b77f57",
+        "#b88059",
+        "#b8825b",
+        "#b9835d",
+        "#ba855f",
+        "#bb8661",
+        "#bb8863",
+        "#bc8965",
+        "#bd8a67",
+        "#bd8c69",
+        "#be8d6b",
+        "#bf8f6c",
+        "#c0906e",
+        "#c09270",
+        "#c19372",
+        "#c29474",
+        "#c29676",
+        "#c39778",
+        "#c4997a",
+        "#c59a7c",
+        "#c59c7e",
+        "#c69d80",
+        "#c79f82",
+        "#c7a084",
+        "#c8a286",
+        "#c9a388",
+        "#caa48a",
+        "#caa68c",
+        "#cba78e",
+        "#cca990",
+        "#cdaa92",
+        "#cdac94",
+        "#cead96",
+        "#cfaf98",
+        "#d0b09a",
+        "#d0b29c",
+        "#d1b39e",
+        "#d2b5a0",
+        "#d3b6a2",
+        "#d3b8a4",
+        "#d4b9a6",
+        "#d5bba8",
+        "#d6bcaa",
+        "#d6beac",
+        "#d7bfae",
+        "#d8c1b0",
+        "#d9c2b2",
+        "#dac4b4",
+        "#dac6b6",
+        "#dbc7b8",
+        "#dcc9ba",
+        "#ddcabc",
+        "#deccbe",
+        "#dfcdc1",
+        "#dfcfc3",
+        "#e0d0c5",
+        "#e1d2c7",
+        "#e2d4c9",
+        "#e3d5cb",
+        "#e4d7cd",
+        "#e5d8cf",
+        "#e5dad1",
+        "#e6dcd3",
+        "#e7ddd5",
+        "#e8dfd7",
+        "#e9e0d9",
+        "#eae2db",
+        "#ebe4dd",
+        "#ece5df",
+        "#ede7e2",
+        "#eee8e4",
+        "#efeae6",
+        "#f0ece8",
+        "#f1edea",
+        "#f2efec",
+        "#f4f0ee",
+        "#f5f2f0",
+        "#f6f4f2",
+        "#f7f5f4",
+        "#f9f7f6",
+        "#faf8f7",
+        "#fbfaf9",
+        "#fcfcfb",
+        "#fefdfd",
+        "#FFFFFF",
+        "#fdfdfd",
+        "#fbfbfb",
+        "#f8f9f9",
+        "#f6f8f7",
+        "#f4f6f6",
+        "#f2f4f4",
+        "#f0f2f2",
+        "#eef0f0",
+        "#ebeeee",
+        "#e9edec",
+        "#e7ebeb",
+        "#e5e9e9",
+        "#e3e7e7",
+        "#e1e6e5",
+        "#dfe4e3",
+        "#dde2e2",
+        "#dbe0e0",
+        "#d9dfde",
+        "#d7dddc",
+        "#d5dbdb",
+        "#d3dad9",
+        "#d1d8d7",
+        "#cfd6d6",
+        "#cdd5d4",
+        "#cbd3d2",
+        "#c9d1d1",
+        "#c7d0cf",
+        "#c5cecd",
+        "#c3cccc",
+        "#c1cbca",
+        "#bfc9c8",
+        "#bdc8c7",
+        "#bbc6c5",
+        "#b9c5c3",
+        "#b7c3c2",
+        "#b5c1c0",
+        "#b3c0bf",
+        "#b1bebd",
+        "#afbdbb",
+        "#adbbba",
+        "#acbab8",
+        "#aab8b7",
+        "#a8b7b5",
+        "#a6b5b4",
+        "#a4b4b2",
+        "#a2b2b1",
+        "#a0b1af",
+        "#9eafae",
+        "#9daeac",
+        "#9bacab",
+        "#99aba9",
+        "#97a9a8",
+        "#95a8a6",
+        "#93a7a5",
+        "#92a5a3",
+        "#90a4a2",
+        "#8ea2a0",
+        "#8ca19f",
+        "#8a9f9d",
+        "#899e9c",
+        "#879d9b",
+        "#859b99",
+        "#839a98",
+        "#819896",
+        "#809795",
+        "#7e9693",
+        "#7c9492",
+        "#7a9391",
+        "#79928f",
+        "#77908e",
+        "#758f8c",
+        "#738e8b",
+        "#728c8a",
+        "#708b88",
+        "#6e8a87",
+        "#6c8886",
+        "#6b8784",
+        "#698683",
+        "#678482",
+        "#658380",
+        "#64827f",
+        "#62807d",
+        "#607f7c",
+        "#5e7e7b",
+        "#5d7c79",
+        "#5b7b78",
+        "#597a77",
+        "#577975",
+        "#557774",
+        "#547673",
+        "#527572",
+        "#507370",
+        "#4e726f",
+        "#4d716e",
+        "#4b706c",
+        "#496e6b",
+        "#476d6a",
+        "#456c68",
+        "#446b67",
+        "#426966",
+        "#406865",
+        "#3e6763",
+        "#3c6662",
+        "#3b6461",
+        "#39635f",
+        "#37625e",
+        "#35615d",
+        "#335f5c",
+        "#315e5a",
+        "#2f5d59",
+        "#2d5c58",
+        "#2b5a57",
+        "#295955",
+        "#275854",
+        "#255753",
+        "#225552",
+        "#205450",
+        "#1e534f",
+        "#1b524e",
+        "#19504d",
+        "#164f4b",
+        "#134e4a",
+    ],
+}
