--- tidy3d/components/viz/palette.py
@@ -1,110 +1,0 @@
-FLEXCOMPUTE_TINTS = {
-    "green": [
-        "#F0FDFA",
-        "#CCFBF1",
-        "#99F6E4",
-        "#5EEAD4",
-        "#2DD4BF",
-        "#14B8A6",
-        "#0D9488",
-        "#0F766E",
-        "#115E59",
-        "#134E4A",
-    ],
-    "green-deep": [
-        "#E2EEE7",
-        "#BEE6CE",
-        "#93D5AD",
-        "#67C48C",
-        "#3CB36B",
-        "#26AB5B",
-        "#229A52",
-        "#1E8949",
-        "#176737",
-        "#0F4424",
-    ],
-    "blue": [
-        "#E6EFFF",
-        "#CCE0FF",
-        "#99C0FF",
-        "#66A1FF",
-        "#3381FF",
-        "#0062FF",
-        "#0058E6",
-        "#0045B3",
-        "#003180",
-        "#002766",
-    ],
-    "yellow": [
-        "#FFFBEB",
-        "#FEF3C7",
-        "#FDE68A",
-        "#FCD34D",
-        "#FBBF24",
-        "#F59E0B",
-        "#D97706",
-        "#B45309",
-        "#92400E",
-        "#78350F",
-    ],
-    "orange": [
-        "#FFF2E7",
-        "#FFE5CF",
-        "#FFCA9E",
-        "#FFB06E",
-        "#FF8825",
-        "#FF7B0D",
-        "#E66F0C",
-        "#CC620A",
-        "#B35609",
-        "#994A08",
-    ],
-    "red": [
-        "#FFF1F2",
-        "#FFE4E6",
-        "#FECDD3",
-        "#FDA4AF",
-        "#FB7185",
-        "#F44E6A",
-        "#E11D48",
-        "#BE123C",
-        "#9F1239",
-        "#881337",
-    ],
-    "purple": [
-        "#DED2FF",
-        "#C5B2FA",
-        "#A78BF7",
-        "#8A65F5",
-        "#7C51F3",
-        "#6D3EF2",
-        "#6238DA",
-        "#5732C2",
-        "#4C2BA9",
-        "#412591",
-    ],
-    "grey": [
-        "#F4F5F7",
-        "#F0F2F5",
-        "#ECEEF1",
-        "#E0E2E7",
-        "#CDD0D9",
-        "#B4B8C4",
-        "#979BAA",
-        "#777C8D",
-        "#5B5F70",
-        "#434756",
-    ],
-    "primary": [
-        "#E6F5F6",
-        "#CDECEC",
-        "#9BD9D9",
-        "#69C5C7",
-        "#37B2B4",
-        "#059FA1",
-        "#058F91",
-        "#047F81",
-        "#046F71",
-        "#035051",
-    ],
-}
